+++ core/game.h
@@ -4,6 +4,9 @@
 
 #pragma once
 
+#include <SDL2/SDL.h>
+#include <SDL2/SDL_image.h>
+
 #include "dtypes.h"
 #include "log.h"
 #include "scene.h"