+++ core/log.h
@@ -16,7 +16,7 @@ namespace Verse
         void warn(str p_info, ...);
         void error(str p_info, ...);
         void graphics(str p_info, ...);
-        void debug(bool p_debug, str p_info, ...);
+        void debug(str p_info, ...);
     
         void vec2(Vec2 p_vector, str p_name = "");
         void num(int p_number, str p_name = "");